Code update & cleaning

+++ lib/helpers.sh
@@ -96,7 +96,19 @@ handle_build_update() {
     if is_server_running; then
         local current_build=$(get_current_build_number)
         if [[ "$requested_build_number" = "latest" ]]; then
-            check_for_update
+            local latest_build=$(get_latest_build "$(get_latest_version)")
+            if [[ "$latest_build" != "$current_build" ]]; then
+                log_bold_nodate_warning "[WARNING] - Server update available. Current build: #$current_build, Latest build: #$latest_build."
+                if confirm_action "Do you want to update to the latest build?"; then
+                    get_build "$BASE_DIR" "$SERVER_FOLDER"
+                    start_minecraft_server
+                else
+                    offer_to_attach
+                fi
+            else
+                log_bold_nodate_info "Server is already running the latest build #$current_build."
+                offer_to_attach
+            fi
         elif [[ "$requested_build_number" != "$current_build" ]]; then
             log_bold_nodate_warning "Server is currently running build #$current_build. Please stop the server before changing builds."
             check_and_attach_screen_session
@@ -109,6 +121,7 @@ handle_build_update() {
         start_minecraft_server
     fi
 }
+
 
 # Function to offer attaching to the current session
 offer_to_attach() {
@@ -141,10 +154,10 @@ is_server_running() {
 
 # Function specifically for starting the server when called by systemd
 start_server_for_systemd() {
-    log_nodate_info "Starting Minecraft server for systemd..."
+    log_bold_nodate_note "Starting Minecraft server for systemd..."
     get_build "$BASE_DIR" "$SERVER_FOLDER"
     start_minecraft_server
-    log_nodate_info "Minecraft server started by systemd."
+    log_bold_nodate_info "Minecraft server started by systemd."
 }
 
 # Parse command-line arguments
@@ -168,6 +181,14 @@ parse_arguments() {
 
 # Function to stop, disable, and remove the systemd service
 stop_disable_and_remove_systemctl() {
+    local service_file="/etc/systemd/system/$SERVICE_NAME.service"
+
+    # Exit if the service file does not exist
+    if [ ! -f "$service_file" ]; then
+        log_bold_nodate_warning "$SERVICE_NAME service file does not exist. Skipping stop, disable, and remove process."
+        return 0
+    fi
+
     # Check if the service is active and stop it
     if systemctl --quiet is-active $SERVICE_NAME; then
         log_bold_nodate_info "Stopping $SERVICE_NAME service..."
@@ -198,20 +219,24 @@ stop_disable_and_remove_systemctl() {
     fi
 
     # Remove the service file
-    if [ -f "/etc/systemd/system/$SERVICE_NAME.service" ]; then
-        log_bold_nodate_info "Removing $SERVICE_NAME service file..."
-        sudo rm "/etc/systemd/system/$SERVICE_NAME.service"
-        sudo systemctl daemon-reload
-        sudo systemctl reset-failed
-    fi
+    log_bold_nodate_info "Removing $SERVICE_NAME service file..."
+    sudo rm "$service_file"
+    sudo systemctl daemon-reload
+    sudo systemctl reset-failed
 
     log_bold_nodate_success "$SERVICE_NAME.service stopped, disabled, and removed successfully."
 }
+
 
 # Function to create the systemd service
 create_systemctl() {
     local session_name="$SESSION"
     local eula_file="$SERVER_FOLDER/eula.txt"
+
+    # Check and create SERVER_FOLDER if it doesn't exist
+    if [ ! -d "$SERVER_FOLDER" ]; then
+        mkdir -p "$SERVER_FOLDER"
+    fi
 
     # Check if service already exists
     if systemctl --quiet is-active $SERVICE_NAME || systemctl --quiet is-enabled $SERVICE_NAME; then
@@ -221,10 +246,10 @@ create_systemctl() {
 
     # EULA acceptance check
     if [[ ! -f "$eula_file" || $(grep -c 'eula=false' "$eula_file") -gt 0 ]]; then
-        echo "You must accept the Minecraft EULA to start the server."
-        echo "Read it here: https://aka.ms/MinecraftEULA"
+        log_bold_nodate_tip "You must accept the Minecraft EULA to start the server."
+        log_bold_nodate_tip "Read it here: https://aka.ms/MinecraftEULA"
         if confirm_action "Do you accept the Minecraft EULA?"; then
-            echo "eula=true" > "$eula_file"
+            log_bold_nodate_prompt "eula=true" > "$eula_file"
             log_bold_nodate_confirmation "EULA accepted."
         else
             log_bold_nodate_error "Minecraft EULA was declined. Cannot start the server."
@@ -256,7 +281,6 @@ EOF"
 
     log_bold_nodate_success "$SERVICE_NAME.service created and started successfully."
 }
-
 # Function to check and attach to existing screen session
 check_and_attach_screen_session() {
     local session_name="$SESSION"
@@ -284,6 +308,7 @@ check_and_attach_screen_session() {
 get_build() {
     local base_dir=$1
     local server_folder=$2
+    local requested_build_number=$PAPER_VERSION
     local session_name="$SESSION"
     local server_jar="$SERVER_JAR"
     local symlink_path="${base_dir}/${server_folder}/$server_jar"
@@ -294,18 +319,22 @@ get_build() {
         exit 1
     fi
 
-    # Extract the current symlinked build number
-    local current_build=$(get_current_build_number)
-
-    # Determine the requested build number based on PAPER_VERSION
+    # Determine the latest version and build number
     local latest_version=$(get_latest_version)
-    local requested_build_number
-    if [ "$PAPER_VERSION" = "latest" ]; then
-        requested_build_number=$(get_latest_build "$latest_version")
-        log_nodate_info "Fetching latest build: $requested_build_number..."
+    local latest_build=$(get_latest_build "$latest_version")
+
+    # Handle 'latest' version setting
+    if [ "$requested_build_number" = "latest" ]; then
+        requested_build_number=$latest_build
+    fi
+
+    # Check for updates if running version is not the requested version
+    local current_build=$(get_current_build_number)
+    if [[ "$requested_build_number" != "$current_build" ]]; then
+        log_bold_nodate_info "Fetching build: $requested_build_number (Current build: $current_build)..."
     else
-        requested_build_number="$PAPER_VERSION"
-        log_nodate_info "Fetching specified build: $requested_build_number..."
+        log_bold_nodate_info "Current build ($current_build) matches the requested build. No need to update."
+        return 0
     fi
 
     local file_name="paper-$latest_version-$requested_build_number.jar"
@@ -314,23 +343,21 @@ get_build() {
     mkdir -p "${base_dir}/lib/builds"
     mkdir -p "${base_dir}/${server_folder}"
 
+    # Download build if it doesn't exist
     if [ ! -f "$file_path" ]; then
-        log_nodate_info "Downloading $file_name ..."
-        if curl -o "$file_path" -# "https://api.papermc.io/v2/projects/paper/versions/$latest_version/builds/$requested_build_number/downloads/$file_name"; then
-            log_bold_nodate_success "Download complete."
-            log_bold_info "Downloaded $file_name."
-        else
+        log_bold_nodate_info "Downloading $file_name ..."
+        if ! curl -o "$file_path" -# "https://api.papermc.io/v2/projects/paper/versions/$latest_version/builds/$requested_build_number/downloads/$file_name"; then
             log_bold_nodate_error "Download failed. Please check the build number and version, and try again."
             exit 1
         fi
+        log_bold_nodate_success "Download complete. Downloaded $file_name."
     else
         log_nodate_note "$file_name already exists. Skipping download."
     fi
 
-    if [ "$requested_build_number" != "$current_build" ] || [ ! -L "$symlink_path" ]; then
-        if [ -L "$symlink_path" ] || [ -f "$symlink_path" ]; then
-            rm "$symlink_path"
-        fi
+    # Link build if different from current build
+    if [ ! -L "$symlink_path" ] || [ "$(readlink "$symlink_path")" != "$file_path" ]; then
+        [ -L "$symlink_path" ] && rm "$symlink_path"
         ln -s "$file_path" "$symlink_path"
         log_nodate_note "Linked $file_name to ${server_folder}/$server_jar."
     else
@@ -356,7 +383,7 @@ configure_minecraft_server() {
         echo "server-port=$SERVER_PORT"
     } > "$server_properties"
 
-    log_nodate_success "Minecraft server properties configured."
+    log_bold_nodate_note "Minecraft server properties configured."
 }
 
 
@@ -391,7 +418,7 @@ start_minecraft_server() {
 
     # Check if EULA is already accepted
     if [[ -f "$eula_file" && $(grep -c 'eula=true' "$eula_file") -gt 0 ]]; then
-        log_nodate_success "EULA already accepted. Starting the server..."
+        log_bold_nodate_note "EULA already accepted. Starting the server..."
     else
         # Prompt for EULA acceptance
         log_bold_nodate_prompt "Do you accept the Minecraft EULA? (https://aka.ms/MinecraftEULA) [y/N]"
@@ -401,14 +428,14 @@ start_minecraft_server() {
             echo "#By changing the setting below to TRUE you are indicating your agreement to our EULA (https://aka.ms/MinecraftEULA)." > "$eula_file"
             echo "#$(date) - grab current timestamp" >> "$eula_file"
             echo "eula=true" >> "$eula_file"
-            log_nodate_success "EULA accepted. Starting the server..."
+            log_bold_nodate_confirmation "EULA accepted. Starting the server..."
         else
             # User declined the EULA, exit the script
-            log_bold_nodate_error "Minecraft EULA was declined. Exiting script."
+            log_bold_nodate_error "Minecraft EULA was declined. Exiting."
             exit 1
         fi
     fi
-    
+
     # Start the server in a detached screen session and redirect output to logfile
     (
         configure_minecraft_server
